refactor-app: 여행 생성 패키지 구조 정리
