+++ application/lib/common/router/router_provider.dart
@@ -1,10 +1,10 @@
 import 'package:application_new/common/session/session_provider.dart';
 import 'package:application_new/feature/authentication/page/login_page.dart';
 import 'package:application_new/feature/home/home_page.dart';
-import 'package:application_new/feature/travel/create_travel_page.dart';
-import 'package:application_new/feature/travel/select_travel_city_form.dart';
-import 'package:application_new/feature/travel/select_travel_detail_form.dart';
-import 'package:application_new/feature/travel/select_travel_date_form.dart';
+import 'package:application_new/feature/travel/page/create_travel_page.dart';
+import 'package:application_new/feature/travel/page/select_travel_city_form.dart';
+import 'package:application_new/feature/travel/page/select_travel_detail_form.dart';
+import 'package:application_new/feature/travel/page/select_travel_date_form.dart';
 import 'package:go_router/go_router.dart';
 
 import 'package:riverpod_annotation/riverpod_annotation.dart';
