+++ application/lib/feature/travel/page/select_travel_city_form.dart
@@ -2,7 +2,7 @@ import 'package:application_new/common/util/translation.dart';
 import 'package:application_new/feature/region/model/region_model.dart';
 import 'package:application_new/feature/region/provider/region_provider.dart';
 import 'package:application_new/feature/travel/component/bottom_action_button.dart';
-import 'package:application_new/feature/travel/create_travel_provider.dart';
+import 'package:application_new/feature/travel/provider/create_travel_provider.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
